refactoring: move makeHash.js to the hash folder
